+++ tests/mock/stub.js
@@ -1,0 +1,4 @@
+function stub() {
+}
+
+export default stub